e2e Test Parsed Request Handler on Sell ETH transaction
